Added logger.c and logger.h
CONTENT: 
-logger.c : the actual class for the project
-loggerM.c : the testing class with a main
-client.c : class to use to send messages to loggerM.c

HOW TO TEST:
compile loggerM.c and client.c. Execute them in 2 different shells, starting with loggerM.c.
In client.c, you'll be prompted a to input a number, input a number>1 to keep going, 1 to end the message queue.
For safety loggerM.c has a built in system to delete all the msgqueues after 2 minutes.